requirements.txt
django
django-model-utils
django-categories
django-taggit

markdown